+++ spec/helpers/fdk.js
@@ -2,19 +2,25 @@ const { setupFdk } = require("../../express");
 const { RedisStorage } = require("../../express/storage");
 const { redisConnection } = require("../helpers/setup_db");
 
-module.exports = (settings) => {
-    return setupFdk({
-        api_key: "API_KEY",
-        api_secret: "API_SECRET",
-        base_url: "http://localdev.fyndx0.de",
-        scopes: ["company/products"],
-        callbacks: {
-            auth: ()=>{},
-            uninstall: ()=>{}
-        },
-        storage: new RedisStorage(redisConnection, "test_fdk"),
-        access_mode: "online",
-        cluster: "http://localdev.fyndx0.de",
-        ...settings
-    })
+const fdkConfig = {
+    api_key: "API_KEY",
+    api_secret: "API_SECRET",
+    base_url: "http://localdev.fyndx0.de",
+    scopes: ["company/products"],
+    callbacks: {
+        auth: ()=>{},
+        uninstall: ()=>{}
+    },
+    storage: new RedisStorage(redisConnection, "test_fdk"),
+    access_mode: "online",
+    cluster: "http://localdev.fyndx0.de",
 };
+
+const initializeFDK = (settings) => {
+    const { setupFdk } = require("../../express");
+    return setupFdk({...fdkConfig, ...settings})
+};
+
+module.exports = {
+    initializeFDK,
+}